requirements.txt
pillow>=10.0.0
instagrapi>=2.0.0
pandas>=2.0.0